fxD3D M3: host display-driver skeleton + INF + agent deploy
M3 (host chassis that loads/registers fxD3D) and the fleet deploy path:

- driver/ddi_glue.[ch]: the portable bridge the driver's DrawPrimitives2 callback
  reduces to (fxdglue_draw_primitives2 -> fxd_dp2_execute). Cross-compiles for
  Win32 and PASSES a host glue test (context + DrawPrimitives2 path, stub backend,
  no DDK/hardware).
- driver/fxd3d.inf: registration - binds PCI VEN_121A DEV 0005 (V3)/0009 (V4/5)
  under the Display class, copies files, writes Display\0000 (applied via PnP).
- driver/nt/ (enable.c + SOURCES + .def): Windows 2000/XP GDI display-driver host
  with the real DDI entry points (DrvGetDirectDrawInfo + D3D ContextCreate/
  DrawPrimitives2) wiring fxD3D in; builds under the NT DDK.
- driver/win9x/d3dcb.c: Win98/ME DDHAL D3D-callback registration
  (GetDriverInfo(GUID_D3DCallbacks3) -> DrawPrimitives2 -> glue) atop a
  vmdisp9x-shaped minivdd; builds under the 98 DDK.
- push_gfxbench.py: deploy gfxbench.exe (+ optional rebuilt glide3x.dll) to a
  fleet Voodoo box over the agent, run the headless sweep, pull the CSV back.
- scripts/3dfx/{README.md,Makefile}: one-command build of all DDK-independent
  parts + both host tests; ties the whole project together.

Fix: DWORD mirror was 'unsigned long' (64-bit on the Linux host) -> now
'unsigned int' to match the Win32 ABI, keeping the packed DP2 layout identical
on host and target (was corrupting struct sizes in the glue test).

Full tree builds green from  in scripts/3dfx: Glide DLLs (h3/h5/glide2) +
gfxbench.exe + fxD3D core + driver glue, with both host unit tests passing.
Remaining for M4/M5: build the host driver in the platform DDK, bring up in
86Box, then real cards.

+++ scripts/3dfx/d3dhal/include/fxd3d.h
@@ -24,7 +24,10 @@
 #  include <d3dhal.h>
 #else
 /* ---- minimal public DDI mirrors (documented MS shapes) ------------------- */
-typedef unsigned long  DWORD;
+/* Match the Windows ABI exactly: DWORD is 32-bit on Win32 (LLP64), NOT the
+ * host's 64-bit `long`. Using fixed-width here keeps the packed DP2 layout
+ * identical on the 64-bit build host (unit tests) and the 32-bit target. */
+typedef unsigned int   DWORD;   /* 32-bit on both host and Win32             */
 typedef unsigned short WORD;
 typedef unsigned char  BYTE;
 typedef float          D3DVALUE;
